refactor de componentens labelDatosUsuarios y labelPagos
