+++ src/app/components/installButton/InstallButton.tsx
