+++ src/app/components/mensajes/labelMensajes.tsx
@@ -1,0 +1,14 @@
+"use Client";
+import { useState, useEffect } from "react";
+
+const LabelMensajes = () => {
+
+
+return (
+    <div className="text-white p-4">
+      <h1 className="text-xl font-bold mb-4">Mensajes del Usuario</h1>
+      <p>Hola soy el label de mensajes</p>
+    </div>
+)
+}
+export default LabelMensajes;
