+++ src/interface/pago.ts
@@ -1,0 +1,6 @@
+export interface  Pago {
+    id_mes: number;
+    created_at: string;
+    year: number;
+    monto: number;
+  }
